requirements.txt
requests==2.31.0
python-telegram-bot==20.7
python-dotenv==1.0.0
streamlit==1.29.0
schedule==1.2.1
pandas==2.1.4
